Move utils to model folder
