+++ src/transformers/models/mask_rcnn/configuration_maskrcnn.py
@@ -36,7 +36,7 @@ class MaskRCNNConfig(PretrainedConfig):
     This is the configuration class to store the configuration of a [`MaskRCNNForObjectDetection`]. It is used to
     instantiate a Mask R-CNN model according to the specified arguments, defining the model architecture. Instantiating
     a configuration with the defaults will yield a similar configuration to that of the Mask R-CNN
-    [microsoft/convnext-tiny-maskrcnn](https://huggingface.co/facebook/convnext-tiny-maskrcnn) architecture.
+    [nielsr/convnext-tiny-maskrcnn](https://huggingface.co/nielsr/convnext-tiny-maskrcnn) architecture.
 
     Configuration objects inherit from [`PretrainedConfig`] and can be used to control the model outputs. Read the
     documentation from [`PretrainedConfig`] for more information.
@@ -53,6 +53,7 @@ class MaskRCNNConfig(PretrainedConfig):
         (...)
 
     Example:
+
     ```python
     >>> from transformers import MaskRCNNConfig, MaskRCNNForObjectDetection
