+++ src/transformers/models/mask_rcnn/image_processing_maskrcnn.py
@@ -1,5 +1,5 @@
 # coding=utf-8
-# Copyright 2022 The HuggingFace Inc. team. All rights reserved.
+# Copyright 2023 The HuggingFace Inc. team. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -43,7 +43,7 @@ if is_torch_available():
     from .modeling_maskrcnn import MaskRCNNDeltaXYWHBBoxCoder
 
 if is_torchvision_available():
-    from ...nms import multiclass_nms
+    from transformers.models.mask_rcnn.nms import multiclass_nms
 
 
 logger = logging.get_logger(__name__)
